+++ src/indicators/jrc_agricultural_hotspots/update_agricultural_hotspots.R
@@ -9,9 +9,7 @@ box::use(../../signals/generate_signals)
 box::use(../../utils/hs_logger)
 box::use(../../utils/update_coverage)
 
-first_run <- as.logical(Sys.getenv("FIRST_RUN", unset = FALSE))
-dry_run <- as.logical(Sys.getenv("HS_DRY_RUN", unset = TRUE))
-dry_run_filter <- if (dry_run) c("AFG", "SSD") else NULL
+dry_run_filter <- if (Sys.getenv("HS_DRY_RUN")) c("AFG", "SSD") else NULL
 indicator_id <- "jrc_agricultural_hotspots"
 
 hs_logger$configure_logger()
@@ -34,7 +32,5 @@ df_jrc <- generate_signals$generate_signals(
   summary_fn = summary_agricultural_hotspots$summary,
   info_fn = info_agricultural_hotspots$info,
   plot_fn = plot_agricultural_hotspots$plot,
-  dry_run = dry_run,
-  dry_run_filter = dry_run_filter,
-  first_run = first_run
+  dry_run_filter = dry_run_filter
 )